+++ tests/mocks/util/MockTextObject.php
@@ -8,7 +8,7 @@
 
 namespace lithium\tests\mocks\util;
 
-class MockStringObject extends \lithium\template\view\Renderer {
+class MockTextObject extends \lithium\template\view\Renderer {
 
 	public $message = 'custom object';