+++ util/String.php
@@ -8,13 +8,28 @@
 
 namespace lithium\util;
 
+use lithium\util\Text;
+use lithium\security\Hash;
 use lithium\security\Random;
 
+$message  = "lithium\util\String has been deprecated in favor of ";
+$message .= "lithium\util\Text and lithium\security\{Hash,Random}. ";
+$message .= "The old class and methods continue to work and redirect calls ";
+$message .= "to the new classes. However it is not possible to use the String ";
+$message .= "class with PHP >=7.0.";
+trigger_error($message, E_USER_DEPRECATED);
+
 /**
- * String manipulation utility class. Includes functionality for generating UUIDs,
- * {:tag} and regex replacement, and tokenization.
+ * String manipulation utility class.
+ *
+ * @deprecated
  */
 class String {
+
+	/**
+	 * Option flag used in `String::random()`.
+	 */
+	const ENCODE_BASE_64 = 1;
 
 	/**
 	 * UUID-related constant. Clears all bits of version byte (`00001111`).
@@ -37,119 +52,81 @@ class String {
 	const UUID_VAR_RFC = 128;
 
 	/**
+	 * Generates random bytes for use in UUIDs and password salts, using
+	 * (when available) a cryptographically strong random number generator.
+	 *
+	 * @param integer $bytes The number of random bytes to generate.
+	 * @param array $options
+	 * @return string Returns a string of random bytes.
+	 */
+	public static function random($bytes, array $options = array()) {
+		$message  = "lithium\util\String::random() has been deprecated in favor of ";
+		$message .= "lithium\security\Random::generate().";
+		trigger_error($message, E_USER_DEPRECATED);
+
+		return Random::generate($bytes, $options);
+	}
+
+	/**
+	 * Uses PHP's hashing functions to create a hash of the string provided.
+	 *
+	 * @param string $string The string to hash.
+	 * @param array $options
+	 * @return string Returns a hashed string.
+	 */
+	public static function hash($string, array $options = array()) {
+		$message  = "lithium\util\String::hash() has been deprecated in favor of ";
+		$message .= "lithium\security\Hash::calculate().";
+		trigger_error($message, E_USER_DEPRECATED);
+
+		return Hash::calculate($string, $options);
+	}
+
+	/**
+	 * Compares two strings in constant time to prevent timing attacks.
+	 *
+	 * @param string $known The string of known length to compare against.
+	 * @param string $user The user-supplied string.
+	 * @return boolean Returns a boolean indicating whether the two strings are equal.
+	 */
+	public static function compare($known, $user) {
+		$message  = "lithium\util\String::compare() has been deprecated in favor of ";
+		$message .= "lithium\security\Hash::compare().";
+		trigger_error($message, E_USER_DEPRECATED);
+
+		return Hash::compare($known, $user);
+	}
+
+	/**
 	 * Generates an RFC 4122-compliant version 4 UUID.
 	 *
 	 * @return string The string representation of an RFC 4122-compliant, version 4 UUID.
 	 * @link http://www.ietf.org/rfc/rfc4122.txt RFC 4122: UUID URN Namespace
 	 */
 	public static function uuid() {
-		$uuid = Random::generate(16);
-		$uuid[6] = chr(ord($uuid[6]) & static::UUID_CLEAR_VER | static::UUID_VERSION_4);
-		$uuid[8] = chr(ord($uuid[8]) & static::UUID_CLEAR_VAR | static::UUID_VAR_RFC);
+		$message  = "lithium\util\String::uuid() has been deprecated in favor of ";
+		$message .= "lithium\util\Text::uuid().";
+		trigger_error($message, E_USER_DEPRECATED);
 
-		return join('-', array(
-			bin2hex(substr($uuid, 0, 4)),
-			bin2hex(substr($uuid, 4, 2)),
-			bin2hex(substr($uuid, 6, 2)),
-			bin2hex(substr($uuid, 8, 2)),
-			bin2hex(substr($uuid, 10, 6))
-		));
+		return Text::uuid();
 	}
 
 	/**
 	 * Replaces variable placeholders inside a string with any given data. Each key
 	 * in the `$data` array corresponds to a variable placeholder name in `$str`.
 	 *
-	 * Usage:
-	 * ```
-	 * String::insert(
-	 *     'My name is {:name} and I am {:age} years old.',
-	 *     array('name' => 'Bob', 'age' => '65')
-	 * ); // returns 'My name is Bob and I am 65 years old.'
-	 * ```
-	 *
-	 * Please note that optimization have applied to this method and parts of the code
-	 * may look like it can refactored or removed but in fact this is part of the applied
-	 * optimization. Please check the history for this section of code before refactoring
-	 *
 	 * @param string $str A string containing variable place-holders.
 	 * @param array $data A key, value array where each key stands for a place-holder variable
 	 *                     name to be replaced with value.
-	 * @param array $options Available options are:
-	 *        - `'after'`: The character or string after the name of the variable place-holder
-	 *          (defaults to `}`).
-	 *        - `'before'`: The character or string in front of the name of the variable
-	 *          place-holder (defaults to `'{:'`).
-	 *        - `'clean'`: A boolean or array with instructions for `String::clean()`.
-	 *        - `'escape'`: The character or string used to escape the before character or string
-	 *          (defaults to `'\'`).
-	 *        - `'format'`: A regular expression to use for matching variable place-holders
-	 *          (defaults to `'/(?<!\\)\:%s/'`. Please note that this option takes precedence over
-	 *          all other options except `'clean'`.
+	 * @param array $options
 	 * @return string
 	 */
 	public static function insert($str, array $data, array $options = array()) {
-		$defaults = array(
-			'before' => '{:',
-			'after' => '}',
-			'escape' => null,
-			'format' => null,
-			'clean' => false
-		);
-		$options += $defaults;
-		$format = $options['format'];
+		$message  = "lithium\util\String::insert() has been deprecated in favor of ";
+		$message .= "lithium\util\Text::insert().";
+		trigger_error($message, E_USER_DEPRECATED);
 
-		if ($format === 'regex' || (!$format && $options['escape'])) {
-			$format = sprintf(
-				'/(?<!%s)%s%%s%s/',
-				preg_quote($options['escape'], '/'),
-				str_replace('%', '%%', preg_quote($options['before'], '/')),
-				str_replace('%', '%%', preg_quote($options['after'], '/'))
-			);
-		}
-
-		if (!$format && key($data) !== 0) {
-			$replace = array();
-
-			foreach ($data as $key => $value) {
-				if (!is_scalar($value)) {
-					if (is_object($value) && method_exists($value, '__toString')) {
-						$value = (string) $value;
-					} else {
-						$value = '';
-					}
-				}
-				$replace["{$options['before']}{$key}{$options['after']}"] = $value;
-			}
-			$str = strtr($str, $replace);
-			return $options['clean'] ? static::clean($str, $options) : $str;
-		}
-
-		if (strpos($str, '?') !== false && isset($data[0])) {
-			$offset = 0;
-
-			while (($pos = strpos($str, '?', $offset)) !== false) {
-				$val = array_shift($data);
-				$offset = $pos + strlen($val);
-				$str = substr_replace($str, $val, $pos, 1);
-			}
-			return $options['clean'] ? static::clean($str, $options) : $str;
-		}
-
-		foreach ($data as $key => $value) {
-			if (!$key = sprintf($format, preg_quote($key, '/'))) {
-				continue;
-			}
-			$hash = crc32($key);
-
-			$str = preg_replace($key, $hash, $str);
-			$str = str_replace($hash, $value, $str);
-		}
-
-		if (!isset($options['format']) && isset($options['before'])) {
-			$str = str_replace($options['escape'] . $options['before'], $options['before'], $str);
-		}
-		return $options['clean'] ? static::clean($str, $options) : $str;
+		return Text::insert($str, $data, $options);
 	}
 
 	/**
@@ -158,66 +135,15 @@ class String {
 	 * and unneeded mark-up around place-holders that did not get replaced by `String::insert()`.
 	 *
 	 * @param string $str The string to clean.
-	 * @param array $options Available options are:
-	 *        - `'after'`: characters marking the end of targeted substring.
-	 *        - `'andText'`: (defaults to `true`).
-	 *        - `'before'`: characters marking the start of targeted substring.
-	 *        - `'clean'`: `true` or an array of clean options:
-	 *          - `'gap'`: Regular expression matching gaps.
-	 *          - `'method'`: Either `'text'` or `'html'` (defaults to `'text'`).
-	 *          - `'replacement'`: String to use for cleaned substrings (defaults to `''`).
-	 *          - `'word'`: Regular expression matching words.
+	 * @param array $options
 	 * @return string The cleaned string.
 	 */
 	public static function clean($str, array $options = array()) {
-		if (is_array($options['clean'])) {
-			$clean = $options['clean'];
-		} else {
-			$clean = array(
-				'method' => is_bool($options['clean']) ? 'text' : $options['clean']
-			);
-		}
+		$message  = "lithium\util\String::clean() has been deprecated in favor of ";
+		$message .= "lithium\util\Text::clean().";
+		trigger_error($message, E_USER_DEPRECATED);
 
-		switch ($clean['method']) {
-			case 'text':
-				$clean += array(
-					'word' => '[\w,.]+',
-					'gap' => '[\s]*(?:(?:and|or|,)[\s]*)?',
-					'replacement' => ''
-				);
-				$before = preg_quote($options['before'], '/');
-				$after = preg_quote($options['after'], '/');
-
-				$kleenex = sprintf(
-					'/(%s%s%s%s|%s%s%s%s|%s%s%s%s%s)/',
-					$before, $clean['word'], $after, $clean['gap'],
-					$clean['gap'], $before, $clean['word'], $after,
-					$clean['gap'], $before, $clean['word'], $after, $clean['gap']
-				);
-				$str = preg_replace($kleenex, $clean['replacement'], $str);
-			break;
-			case 'html':
-				$clean += array(
-					'word' => '[\w,.]+',
-					'andText' => true,
-					'replacement' => ''
-				);
-				$kleenex = sprintf(
-					'/[\s]*[a-z]+=(")(%s%s%s[\s]*)+\\1/i',
-					preg_quote($options['before'], '/'),
-					$clean['word'],
-					preg_quote($options['after'], '/')
-				);
-				$str = preg_replace($kleenex, $clean['replacement'], $str);
-
-				if ($clean['andText']) {
-					return static::clean($str, array(
-						'clean' => array('method' => 'text')
-					) + $options);
-				}
-			break;
-		}
-		return $str;
+		return Text::clean($str, $options);
 	}
 
 	/**
@@ -229,10 +155,11 @@ class String {
 	 * @return mixed
 	 */
 	public static function extract($regex, $str, $index = 0) {
-		if (!preg_match($regex, $str, $match)) {
-			return false;
-		}
-		return isset($match[$index]) ? $match[$index] : null;
+		$message  = "lithium\util\String::extract() has been deprecated in favor of ";
+		$message .= "lithium\util\Text::extract().";
+		trigger_error($message, E_USER_DEPRECATED);
+
+		return Text::extract($regex, $str, $index);
 	}
 
 	/**
@@ -241,145 +168,15 @@ class String {
 	 * `$options['rightBound']`.
 	 *
 	 * @param string $data The data to tokenize.
-	 * @param array $options Options to use when tokenizing:
-	 *              -`'separator'` _string_: The token to split the data on.
-	 *              -`'leftBound'` _string_: Left scope-enclosing boundary.
-	 *              -`'rightBound'` _string_: Right scope-enclosing boundary.
+	 * @param array $options
 	 * @return array Returns an array of tokens.
 	 */
 	public static function tokenize($data, array $options = array()) {
-		$options += array('separator' => ',', 'leftBound' => '(', 'rightBound' => ')');
-
-		if (!$data || is_array($data)) {
-			return $data;
-		}
-
-		$depth = 0;
-		$offset = 0;
-		$buffer = '';
-		$results = array();
-		$length = strlen($data);
-		$open = false;
-
-		while ($offset <= $length) {
-			$tmpOffset = -1;
-			$offsets = array(
-				strpos($data, $options['separator'], $offset),
-				strpos($data, $options['leftBound'], $offset),
-				strpos($data, $options['rightBound'], $offset)
-			);
-
-			for ($i = 0; $i < 3; $i++) {
-				if ($offsets[$i] !== false && ($offsets[$i] < $tmpOffset || $tmpOffset === -1)) {
-					$tmpOffset = $offsets[$i];
-				}
-			}
-
-			if ($tmpOffset === -1) {
-				$results[] = $buffer . substr($data, $offset);
-				$offset = $length + 1;
-				continue;
-			}
-			$buffer .= substr($data, $offset, ($tmpOffset - $offset));
-
-			if ($data[$tmpOffset] === $options['separator'] && $depth === 0) {
-				$results[] = $buffer;
-				$buffer = '';
-			} else {
-				$buffer .= $data{$tmpOffset};
-			}
-
-			if ($options['leftBound'] !== $options['rightBound']) {
-				if ($data[$tmpOffset] === $options['leftBound']) {
-					$depth++;
-				}
-				if ($data[$tmpOffset] === $options['rightBound']) {
-					$depth--;
-				}
-				$offset = ++$tmpOffset;
-				continue;
-			}
-
-			if ($data[$tmpOffset] === $options['leftBound']) {
-				($open) ? $depth-- : $depth++;
-				$open = !$open;
-			}
-			$offset = ++$tmpOffset;
-		}
-
-		if (!$results && $buffer) {
-			$results[] = $buffer;
-		}
-		return $results ? array_map('trim', $results) : array();
-	}
-
-	/* Deprecated / BC */
-
-	/**
-	 * Option flag used in `String::random()`.
-	 *
-	 * @deprecated
-	 */
-	const ENCODE_BASE_64 = 1;
-
-	/**
-	 * Generates random bytes for use in UUIDs and password salts, using
-	 * (when available) a cryptographically strong random number generator.
-	 *
-	 * @deprecated
-	 * @param integer $bytes The number of random bytes to generate.
-	 * @param array $options The options used when generating random bytes:
-	 *              - `'encode'` _integer_: If specified, and set to `String::ENCODE_BASE_64`, the
-	 *                resulting value will be base64-encoded, per the notes above.
-	 * @return string Returns a string of random bytes.
-	 */
-	public static function random($bytes, array $options = array()) {
-		$message  = "lithium\util\String::random() has been deprecated in favor of ";
-		$message .= "lithium\security\Random::generate().";
+		$message  = "lithium\util\String::tokenize() has been deprecated in favor of ";
+		$message .= "lithium\util\Text::tokenize().";
 		trigger_error($message, E_USER_DEPRECATED);
 
-		return \lithium\security\Random::generate($bytes, $options);
-	}
-
-	/**
-	 * Uses PHP's hashing functions to create a hash of the string provided, using the options
-	 * specified. The default hash algorithm is SHA-512.
-	 *
-	 * @deprecated
-	 * @param string $string The string to hash.
-	 * @param array $options Supported options:
-	 *        - `'type'` _string_: Any valid hashing algorithm. See the `hash_algos()` function to
-	 *          determine which are available on your system.
-	 *        - `'salt'` _string_: A _salt_ value which, if specified, will be prepended to the
-	 *          string.
-	 *        - `'key'` _string_: If specified `hash_hmac()` will be used to hash the string,
-	 *          instead of `hash()`, with `'key'` being used as the message key.
-	 *        - `'raw'` _boolean_: If `true`, outputs the raw binary result of the hash operation.
-	 *          Defaults to `false`.
-	 * @return string Returns a hashed string.
-	 */
-	public static function hash($string, array $options = array()) {
-		$message  = "lithium\util\String::hash() has been deprecated in favor of ";
-		$message .= "lithium\security\Hash::calculate().";
-		trigger_error($message, E_USER_DEPRECATED);
-
-		return \lithium\security\Hash::calculate($string, $options);
-	}
-
-	/**
-	 * Compares two strings in constant time to prevent timing attacks.
-	 *
-	 * @deprecated
-	 * @param string $known The string of known length to compare against.
-	 * @param string $user The user-supplied string.
-	 * @return boolean Returns a boolean indicating whether the two strings are equal.
-	 */
-	public static function compare($known, $user) {
-		$message  = "lithium\util\String::compare() has been deprecated in favor of ";
-		$message .= "lithium\security\Hash::compare().";
-		trigger_error($message, E_USER_DEPRECATED);
-
-		return \lithium\security\Hash::compare($known, $user);
+		return Text::tokenize($data, $options);
 	}
 }